cpufreq: schedutil: remove redundant code from sugov_next_freq_shared()
The same code is present both within and outside the loop and it doesn't
look like it provides any additional benefit. Remove the special
handling of sg_cpu and let it happen within the loop.

With this change we will do two extra comparisons for the sg_cpu in the
loop, but the loop will do one less comparison for every other CPU in
the policy.

While at it, also remove the excess parameters of sugov_next_freq_shared().

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
(cherry picked from commit 3e69906805e2d61df10741ad4690d282c0292519)
Signed-off-by: Eliminater74 <eliminater74@gmail.com>